backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
joblib
python-multipart
python-dotenv